sightline-server/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sightline-server"
version = "0.1.0"
description = "YOLO-based object detection inference server"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "ultralytics",
    "pillow",
    "python-multipart",
    "facenet-pytorch>=2.5.3",
    "scikit-learn>=1.3.0",
    "torch>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "ruff",
]

[tool.black]
line-length = 120
target-version = ["py38", "py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 120
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "ISC", "PIE", "PYI", "Q", "RET", "SIM", "ARG"]
ignore = ["E501"]
